requirements.txt
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-celery-results==2.4.0
django-redis==5.2.0
django-redis-cache==3.0.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
importlib-metadata==4.13.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.32
psycopg2-binary==2.8.6
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
redis==3.5.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
zipp==3.10.0
